유저 프로덕트 추가 쿼리로 변경

+++ src/components/UserProductAdd/UserProductCard.tsx
@@ -9,28 +9,52 @@ import {
 } from "@/types/productTypeGurards";
 import Link from "next/link";
 import ButtonBasic from "../designs/ButtonBasic";
+import { useMutation } from "@tanstack/react-query";
 
 interface UserProductCardProps {
   userProduct: GetUserProductResponse;
   onDelete: (id: number) => void;
+  userId: number | null;
 }
 
 export default function UserProductCard({
   userProduct,
   onDelete,
+  userId,
 }: UserProductCardProps) {
   const { isModalOpen, openModal, closeModal } = useModal();
+
+  const { mutate: editProductProductMutationFn } = useMutation({
+    mutationFn: () => {
+      return;
+    },
+    onSuccess: () => {
+      return;
+    },
+    onError: () => {
+      return;
+    },
+  });
 
   const handleEdit = (e: React.MouseEvent<HTMLButtonElement>) => {
     e.preventDefault();
     e.stopPropagation();
+
     console.log("수정하기");
+    editProductProductMutationFn()
   };
 
   const handleDelete = (e: React.MouseEvent<HTMLButtonElement>) => {
     e.preventDefault();
     e.stopPropagation();
-    onDelete(userProduct.id);
+
+    // 리뷰가 있는 경우 모달 확인
+    if (userProduct.reviews.some((review) => review.userId === userId)) {
+      openModal();
+    } else {
+      // 리뷰가 없는 경우 즉시 삭제
+      onDelete(userProduct.id);
+    }
   };
 
   // TODO 유저프로덕트에 reviews조회기능 추가되면 - 리뷰 있는 경우 확인모달 띄우기 추가
@@ -38,68 +62,73 @@ export default function UserProductCard({
     const { myOption, displaySize } = userProduct.product;
 
     return (
-      <Link
-        href={`/Mac/${userProduct.product.id}`}
-        className="user-product-card"
-      >
-        {/* TODO 이미지로 대체 */}
-        <div className="w-16 h-16 bg-gray-300" />
+      <>
+        <Link
+          href={`/Mac/${userProduct.product.id}`}
+          className="user-product-card"
+        >
+          {/* TODO 이미지로 대체 */}
+          <div className="w-16 h-16 bg-gray-300" />
 
-        <div className="flex flex-col items-start gap-1 col-span-2">
-          <h3
-            className={`justify-self-start font-bold ${
-              userProduct.status === "SOLD" && "line-through text-light"
-            }`}
-          >
-            {userProduct.product.name}
-          </h3>
+          <div className="flex flex-col items-start gap-1 col-span-2">
+            <h3
+              className={`justify-self-start font-bold ${
+                userProduct.status === "SOLD" && "line-through text-light"
+              }`}
+            >
+              {userProduct.product.name}
+            </h3>
 
-          <div className="flex gap-1 justify-self-center">
-            <p className="justify-self-center text-xs md:text-sm text-light">
-              {myOption?.processor}
-            </p>
-            <p className="justify-self-center text-xs md:text-sm text-light">
-              {displaySize}
-            </p>
+            <div className="flex gap-1 justify-self-center">
+              <p className="justify-self-center text-xs md:text-sm text-light">
+                {myOption?.processor}
+              </p>
+              <p className="justify-self-center text-xs md:text-sm text-light">
+                {displaySize}
+              </p>
+            </div>
           </div>
-        </div>
 
-        <span className="justify-self-center">
-          {userProduct.status !== "SOLD" && (
-            <MonthDiff purchasedAt={userProduct.purchasedAt} />
-          )}
-        </span>
+          <span className="justify-self-center">
+            {userProduct.status !== "SOLD" && (
+              <MonthDiff purchasedAt={userProduct.purchasedAt} />
+            )}
+          </span>
 
-        <span className="w-full flex justify-between md:justify-end items-center">
-          {userProduct.purchasePrice && userProduct.purchasePrice !== 0 ? (
-            <p className="justify-self-end light-p whitespace-nowrap">{`${userProduct.purchasePrice.toLocaleString()}원`}</p>
-          ) : (
-            <p className="justify-self-end light-p whitespace-nowrap">
-              가격 미입력
-            </p>
-          )}
+          <span className="w-full flex justify-between md:justify-end items-center">
+            {userProduct.purchasePrice && userProduct.purchasePrice !== 0 ? (
+              <p className="justify-self-end light-p whitespace-nowrap">{`${userProduct.purchasePrice.toLocaleString()}원`}</p>
+            ) : (
+              <p className="justify-self-end light-p whitespace-nowrap">
+                가격 미입력
+              </p>
+            )}
 
-          <div className="flex gap-3 md:hidden">
+            <div className="flex gap-3 md:hidden">
+              <ButtonBasic text="수정하기" onClick={handleEdit} />
+              <ButtonBasic text="삭제하기" onClick={handleDelete} />
+            </div>
+          </span>
+
+          <div className="ml-10 hidden md:flex flex-col">
             <ButtonBasic text="수정하기" onClick={handleEdit} />
             <ButtonBasic text="삭제하기" onClick={handleDelete} />
           </div>
-        </span>
-
-        <div className="ml-10 hidden md:flex flex-col">
-          <ButtonBasic text="수정하기" onClick={handleEdit} />
-          <ButtonBasic text="삭제하기" onClick={handleDelete} />
-        </div>
+        </Link>
 
         <Modal
           isModalOpen={isModalOpen}
           onClose={closeModal}
-          onConfirm={closeModal}
+          onConfirm={() => {
+            onDelete(userProduct.id);
+            closeModal();
+          }}
           onCancel={() => closeModal()}
           title="정말 삭제할까요?"
           content="이 제품에는 리뷰가 작성되어 있어요. 그래도 삭제하시겠어요?"
           confirmBtnText="삭제하기"
         />
-      </Link>
+      </>
     );
   }
